Moving stuff around according to the review comments
