+++ mat_dp_pipeline/__init__.py
@@ -1,18 +1,11 @@
 __all__ = [
-    "LabelledOutput",
     "PipelineOutput",
     "pipeline",
     "create_sdf",
     "StandardDataFormat",
-    "Year",
     "App",
 ]
 
-from mat_dp_pipeline.application import App
-from mat_dp_pipeline.pipeline import (
-    LabelledOutput,
-    PipelineOutput,
-    create_sdf,
-    pipeline,
-)
-from mat_dp_pipeline.standard_data_format import StandardDataFormat, Year
+from mat_dp_pipeline.pipeline import PipelineOutput, pipeline
+from mat_dp_pipeline.presentation import App
+from mat_dp_pipeline.sdf import StandardDataFormat, create_sdf
