+++ mat_dp_pipeline/pipeline/__init__.py
@@ -1,0 +1,1 @@
+from mat_dp_pipeline.pipeline.pipeline import PipelineOutput, pipeline
