+++ mat_dp_pipeline/presentation/__init__.py
@@ -1,0 +1,1 @@
+from mat_dp_pipeline.presentation.application import App
